refactor: ui/format.lua の impure 関数を ui/inline.lua へ抽出
check_purity を実行したところ ui/format.lua が「Pure format/calculation
functions with no state or vim API side effects」と CLAUDE.md で宣言
されているにも関わらず、`format_comments_for_inline` とその局所
ヘルパー `wrap_line` が `vim.api`/`vim.fn` でウィンドウ runtime state を
読んでいた (6 件の違反)。

これらを新規モジュール `lua/fude/ui/inline.lua` へ抽出し、format.lua
の純粋性を回復する。

- lua/fude/ui/inline.lua: 新規。`wrap_line` (local) と
  `format_comments_for_inline` を移動。`format.normalize_newlines` /
  `format.apply_markdown_highlight_to_line` は require 経由で利用
- lua/fude/ui/format.lua: 上記 2 関数を削除
- lua/fude/ui.lua: facade で `M.format_comments_for_inline` の参照先を
  inline モジュールに変更
- lua/fude/ui/extmarks.lua: lazy require を inline に変更
- CLAUDE.md: ui/format.lua の責務記述から `format_comments_for_inline`
  を削除、ui/inline.lua の責務記述を追加。あわせて元から記述漏れの
  あった `normalize_newlines` も format.lua の関数リストに補正。purity
  enforced by check_purity を明記

公開 API (`ui.format_comments_for_inline`) は不変のため、既存テスト
(18 件) は無変更でパスする。
